Fixed division

+++ hw01/src/ru/sberbank/homework/kashin/main/java/calculator/SimpleCalculator.java
@@ -25,7 +25,7 @@ public class SimpleCalculator {
         return one * two;
     }
 
-    public int division(int one, int two) {
+    public double division(int one, int two) {
         return one / two;
     }
